+++ README.md
@@ -1,3 +1,6 @@
+### CloudFront URL: https://dg6kd2afb2dhm.cloudfront.net/
+### S3 Bucket URL: http://velo-market.s3-website-eu-west-1.amazonaws.com/
+
 This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
 ## Available Scripts
